+++ scripts/Install-LocalBuild.ps1
@@ -1,7 +1,7 @@
 [CmdletBinding()]
 param(
     [Parameter(Mandatory)][string]$Target,
-    [string]$Version = '1.0.0-rc.3'
+    [string]$Version = '1.0.0-rc.4'
 )
 
 $ErrorActionPreference = 'Stop'